BottomNavigationView:  - Added icons to bottomnavigationview
