+++ app/src/main/res/drawable/ic_line_chart.xml
@@ -1,0 +1,21 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="486.742dp"
+    android:height="486.742dp"
+    android:viewportWidth="486.742"
+    android:viewportHeight="486.742">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M33,362.371v78.9c0,4.8 3.9,8.8 8.8,8.8h61c4.8,0 8.8,-3.9 8.8,-8.8v-138.8l-44.3,44.3C57.9,356.071 45.9,361.471 33,362.371z"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M142,301.471v139.8c0,4.8 3.9,8.8 8.8,8.8h61c4.8,0 8.8,-3.9 8.8,-8.8v-82.3c-13.9,-0.3 -26.9,-5.8 -36.7,-15.6L142,301.471z"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M251,350.271v91c0,4.8 3.9,8.8 8.8,8.8h61c4.8,0 8.8,-3.9 8.8,-8.8v-167.9l-69.9,69.9C257,345.971 254.1,348.271 251,350.271z"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M432.7,170.171l-72.7,72.7v198.4c0,4.8 3.9,8.8 8.8,8.8h61c4.8,0 8.8,-3.9 8.8,-8.8v-265.6c-2,-1.7 -3.5,-3.2 -4.6,-4.2L432.7,170.171z"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M482.6,41.371c-2.9,-3.1 -7.3,-4.7 -12.9,-4.7c-0.5,0 -1.1,0 -1.6,0c-28.4,1.3 -56.7,2.7 -85.1,4c-3.8,0.2 -9,0.4 -13.1,4.5c-1.3,1.3 -2.3,2.8 -3.1,4.6c-4.2,9.1 1.7,15 4.5,17.8l7.1,7.2c4.9,5 9.9,10 14.9,14.9l-171.6,171.7l-77.1,-77.1c-4.6,-4.6 -10.8,-7.2 -17.4,-7.2c-6.6,0 -12.7,2.6 -17.3,7.2L7.2,286.871c-9.6,9.6 -9.6,25.1 0,34.7l4.6,4.6c4.6,4.6 10.8,7.2 17.4,7.2s12.7,-2.6 17.3,-7.2l80.7,-80.7l77.1,77.1c4.6,4.6 10.8,7.2 17.4,7.2c6.6,0 12.7,-2.6 17.4,-7.2l193.6,-193.6l21.9,21.8c2.6,2.6 6.2,6.2 11.7,6.2c2.3,0 4.6,-0.6 7,-1.9c1.6,-0.9 3,-1.9 4.2,-3.1c4.3,-4.3 5.1,-9.8 5.3,-14.1c0.8,-18.4 1.7,-36.8 2.6,-55.3l1.3,-27.7C487,49.071 485.7,44.571 482.6,41.371z"/>
+</vector>
