+++ app/src/main/res/drawable/ic_money_exchange.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="394.653dp"
+    android:height="394.653dp"
+    android:viewportWidth="394.653"
+    android:viewportHeight="394.653">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M180.256,250.992c-9.596,-2.584 -18.186,-7.312 -24.188,-13.31c-2.598,-2.593 -6.051,-4.021 -9.723,-4.021c-3.672,0 -7.125,1.431 -9.725,4.026l-1.051,1.05c-2.59,2.587 -4.059,6.173 -4.033,9.84c0.029,3.661 1.555,7.224 4.188,9.774c11.713,11.346 28.258,19.549 45.393,22.504l1.66,0.287v12.987c0,7.584 6.17,13.756 13.756,13.756h1.559c7.584,0 13.754,-6.172 13.754,-13.756V280.84l1.627,-0.309c28.441,-5.415 45.424,-23.191 45.424,-47.533c0.248,-33.103 -24.271,-45.187 -45.656,-52.016l-1.395,-0.444V135.95l2.43,0.531c6.305,1.379 11.934,3.486 16.279,6.096c2.141,1.282 4.582,1.959 7.064,1.959c4.781,0 9.145,-2.417 11.676,-6.467l0.697,-1.117c1.938,-3.102 2.562,-6.92 1.715,-10.473c-0.848,-3.559 -3.129,-6.687 -6.258,-8.581c-8.768,-5.305 -20.098,-9.222 -31.906,-11.028l-1.697,-0.26v-9.82c0,-7.585 -6.17,-13.757 -13.754,-13.757h-1.559c-7.586,0 -13.756,6.171 -13.756,13.757v10.548l-1.605,0.324c-26.283,5.301 -41.975,21.791 -41.975,44.109c0,23.889 13.811,40.182 42.223,49.812l1.357,0.46v49.626L180.256,250.992zM211.846,211.33l2.807,1.229c11.146,4.891 15.262,10.375 15.176,20.263c0,8.235 -5.166,14.107 -15.355,17.456l-2.625,0.86L211.846,211.33L211.846,211.33zM179.833,169.266c-8.326,-4.448 -11.57,-9.353 -11.57,-17.494c0,-6.188 3.979,-10.792 11.822,-13.684l2.691,-0.993v33.742L179.833,169.266z"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M389.834,75.858c-4.055,-3.244 -9.664,-3.726 -14.215,-1.221l-27.025,14.882C314.969,37.853 257.531,6.327 195.217,6.327c-100.906,0 -183,82.093 -183,183c0,11.046 8.955,20 20,20c11.047,0 20,-8.954 20,-20c0,-78.851 64.15,-143 143,-143c47.693,0 91.748,23.629 118.234,62.543l-25.314,13.939c-4.549,2.505 -7.141,7.503 -6.564,12.665c0.571,5.161 4.199,9.469 9.188,10.914l67.828,19.655c1.172,0.34 2.373,0.507 3.572,0.507c2.146,0 4.279,-0.538 6.192,-1.592c2.981,-1.643 5.19,-4.403 6.14,-7.673l19.654,-67.828C395.59,84.469 393.889,79.102 389.834,75.858z"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M362.436,185.327c-11.047,0 -20,8.954 -20,20c0,78.852 -64.149,143 -143.001,143c-47.692,0 -91.748,-23.629 -118.233,-62.543l25.314,-13.939c4.549,-2.504 7.141,-7.502 6.566,-12.664c-0.572,-5.161 -4.199,-9.47 -9.188,-10.913l-67.827,-19.655c-1.172,-0.341 -2.373,-0.507 -3.572,-0.507c-2.146,0 -4.279,0.537 -6.193,1.592c-2.982,1.643 -5.191,4.402 -6.139,7.673L0.508,305.198c-1.445,4.987 0.256,10.354 4.311,13.599c4.055,3.244 9.664,3.728 14.215,1.222l27.025,-14.882c33.624,51.664 91.062,83.19 153.375,83.19c100.906,0 183.001,-82.093 183.001,-183C382.436,194.281 373.48,185.327 362.436,185.327z"/>
+</vector>
